Implement sample runs.
Former-commit-id: c8fc6609c8f2e45614f88d70188648fcea9dadbb [formerly 91d1d33cdc9f51c99bb6b164fa79f3ab37588bba]

Former-commit-id: 320ab5e640e48b390b012ff4a5c8121ec2596a0b

+++ src/ca/ubc/cs/beta/stationpacking/modelcount/IModelCountSolver.java
@@ -19,7 +19,7 @@ public interface IModelCountSolver {
      * @param aSeed - the execution seed.
      * @return
      */
-    public Long countSatisfiablePackings(StationPackingInstance aInstance, long aSeed);
+    public Double countSatisfiablePackings(StationPackingInstance aInstance, long aSeed);
     
     /**
      * Produces a count (or estimate) of the total number of SAT results involving station subsets
@@ -29,7 +29,7 @@ public interface IModelCountSolver {
      * @param aSeed
      * @return
      */
-    public Long countSatisfiablePackingsContainingStation(StationPackingInstance aInstance, Station aStation, long aSeed);
+    public Double countSatisfiablePackingsContainingStation(StationPackingInstance aInstance, Station aStation, long aSeed);
     
     /**
      * Tries to stop the solve call if implemented, if not throws an UnsupportedOperationException.
